fix: split Resolutions subpage to multiple components
